Jenny/Jenny/YOLO/qc-edge-infer/requirements.txt
ultralytics
opencv-python
fastapi
uvicorn[standard]
numpy
pillow
